[Refactoring] 디렉토리 이름변경
