+++ gg-agenda-api/src/main/java/gg/agenda/GgAgendaApiApplication.java
@@ -1,4 +1,4 @@
-package gg.ggagendaapi;
+package gg.agenda;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
