+++ gg-agenda-api/src/test/java/gg/agenda/GgAgendaApiApplicationTests.java
@@ -1,4 +1,4 @@
-package gg.ggagendaapi;
+package gg.agenda;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
